Done app !

+++ handlers/searchHandler.js
@@ -1,6 +1,6 @@
 const REPLACE_ME_STR = '<div class="replaceMe"></div>';
 
-let queryDataMap = [];
+let queryDataMapReal = {};
 
 const Image = require('../models/ImageSchema');
 const Tag = require('../models/TagSchema');
@@ -9,7 +9,7 @@ const baseHandler = require('../handlers/baseHandler');
 const fs = require('fs');
 
 function parseSearchRequestData(req) {
-    queryDataMap = [];
+    queryDataMapReal = {};
     let queryDataStr = req.url.substr(req.url.indexOf('?') + 1);
     let splitterByAmpersand = queryDataStr.split('&')
 
@@ -17,10 +17,7 @@ function parseSearchRequestData(req) {
         let keyValuePair = pair.split('=');
 
         if (keyValuePair[1] !== '') {
-            queryDataMap.push({
-                key: keyValuePair[0],
-                value: keyValuePair[1]
-            });
+            queryDataMapReal[keyValuePair[0]] = keyValuePair[1]
         }
     }
 }
@@ -33,24 +30,74 @@ function listAll(req, res) {
         }
 
         Image.find({}, function (err, images) {
-            let htmlContent = '';
-            for (let image of images) {
-                let imageInstanceReplacement = `<fieldset><legend id="${image.title}"></legend> 
-                                        <img src="${image.url}">
-                                        </img><p>${image.title}</p>
-                                        </img><p>${image.description}</p>
-                                        <button onclick='location.href="/delete?id=${image._id}"'class='deleteBtn'>Delete</button> 
-                                        </fieldset>`
-                htmlContent += imageInstanceReplacement
-            }
-
-            data = data.toString().replace(REPLACE_ME_STR, htmlContent)
-
-            baseHandler.handleOk(req, res, data)
+            sendImagesToClient(req, res, data, images);
         });
     })
 
 }
+
+function renderImages(images) {
+    let htmlContent = '';
+
+    for (let image of images) {
+        if (image !== null) {
+            htmlContent += `<fieldset><legend id="${image.title}"></legend>
+                                                            <img src="${image.url}">
+                                                            </img><p>${image.title}</p>
+                                                            </img><p>${image.description}</p>
+                                                            <button onclick='location.href="/delete?id=${image._id}"'class='deleteBtn'>Delete</button>
+                                                            </fieldset>`
+        }
+    }
+
+    return htmlContent;
+}
+function sendImagesToClient(req, res, data, images) {
+    let htmlContent = renderImages(images);
+
+    data = data.toString().replace(REPLACE_ME_STR, htmlContent)
+    baseHandler.handleOk(req, res, data)
+}
+
+function listByAllParams(req, res, data) {
+    let tagNameValue = queryDataMapReal['tagName'];
+    let afterDate = queryDataMapReal['afterDate'] || new Date(2000, 1, 1, 0, 0, 0, 0);
+    let beforeDate = queryDataMapReal['beforeDate'] || new Date(2100, 1, 1, 0, 0, 0, 0);
+    let limit = parseInt(queryDataMapReal['Limit']) || 10;
+
+    Tag.findOne({name: tagNameValue}, function (err, tag) {
+        if (tag === null) {
+            baseHandler.handleOk(req, res, data)
+            return
+        }
+
+        Image
+            .find({tags: tag._id})
+            .where('creationDate').gt(afterDate).lt(beforeDate)
+            .limit(limit)
+            .exec((err, images) => {
+                if (err) console.log(err);
+
+                sendImagesToClient(req, res, data, images);
+            })
+    });
+}
+function listByDatesOnly(req, res, data) {
+    let afterDate = queryDataMapReal['afterDate'] || new Date(2000, 1, 1, 0, 0, 0, 0);
+    let beforeDate = queryDataMapReal['beforeDate'] || new Date(2100, 1, 1, 0, 0, 0, 0);
+    let limit = parseInt(queryDataMapReal['Limit']) || 10;
+
+    Image
+        .find({})
+        .where('creationDate').gt(afterDate).lt(beforeDate)
+        .limit(limit)
+        .exec((err, images) => {
+            if (err) console.log(err);
+
+            sendImagesToClient(req, res, data, images);
+        })
+}
+
 function listByParams(req, res) {
     fs.readFile(__dirname + '/../views/results.html', (err, data) => {
         if (err) {
@@ -58,44 +105,18 @@ function listByParams(req, res) {
             return;
         }
 
-        let tagNameValue = queryDataMap[0].value;
-
-        Tag.findOne({name: tagNameValue}, function (err, tag) {
-            if (tag === null) {
-                baseHandler.handleOk(req, res, data)
-                return
-            }
-            let htmlContent = '';
-
-            for (let imageId of tag.images) {
-                Image.findOne({_id: imageId}, function (err, image) {
-                    if (err) console.log(err);
-
-                    if (image !== null) {
-                        let imageInstanceReplacement = `<fieldset><legend id="${image.title}"></legend> 
-                                        <img src="${image.url}">
-                                        </img><p>${image.title}</p>
-                                        </img><p>${image.description}</p>
-                                        <button onclick='location.href="/delete?id=${image._id}"'class='deleteBtn'>Delete</button> 
-                                        </fieldset>`
-                        htmlContent += imageInstanceReplacement
-
-                        data = data.toString().replace(REPLACE_ME_STR, imageInstanceReplacement)
-
-                        baseHandler.handleOk(req, res, data)
-                    }
-                })
-            }
-
-        });
+        if (queryDataMapReal['tagName']) {
+            listByAllParams(req, res, data);
+        } else {
+            listByDatesOnly(req, res, data)
+        }
     })
-
-
 }
+
 function search(req, res) {
     parseSearchRequestData(req);
 
-    if (isEmpty(queryDataMap)) {
+    if (getDictionarySize(queryDataMapReal) === 0) {
         listAll(req, res);
     } else {
         listByParams(req, res)
@@ -103,8 +124,8 @@ function search(req, res) {
 
 }
 
-function isEmpty(obj) {
-    return Object.keys(obj).length === 0;
+function getDictionarySize(obj) {
+    return Object.keys(obj).length;
 }
 
 module.exports = (req, res) => {
